feat: separate cr and tdl tests into separate files
